delete websocket
