+++ internal/messages/usecase/messages_usecase_interface.go
@@ -14,6 +14,5 @@ type MessageUsecase interface {
 	GetMessagesWithPage(ctx context.Context, userId uuid.UUID, chatId uuid.UUID, lastMessageId uuid.UUID) (models.MessagesArrayDTO, error)
 
 	GetMessages(ctx context.Context, chatId uuid.UUID) (models.MessagesArrayDTO, error)
-	ScanForNewMessages(ctx context.Context, channel chan<- models.WebScoketDTO, res chan<- error, closeChannel <-chan bool)
-	GetOnlineUsers() map[uuid.UUID]bool
+
 }
